requirements.txt
pdfplumber
pytesseract
Pillow
pdf2image
python-dateutil
